requirements.txt
cassandra-driver==3.11.0
ccm==2.8.1
certifi==2017.7.27.1
chardet==3.0.4
fix-yahoo-finance==0.0.18
futures==3.1.1
idna==2.6
multitasking==0.0.4
numpy==1.13.1
pandas==0.20.3
pandas-datareader==0.5.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
six==1.10.0
urllib3==1.22
